Fix up Regexp creation for Ruby >= 2.2
Regular expressions and grep were changed somewhere around Ruby 2.2 and
it is no longer possible to simply read a string in from the config file
and use it as a regular expression.  You must now read it in, if it
isn't empty, then you must convert it from a string to a regular
expression, and only then can it be used by the grep calls elsewhere.
This fixes the problem where valid grep strings in the config file
failed to result in grep matches.

Signed-off-by: Doug Ledford <dledford@redhat.com>