Add support for mirror repositories
When using mirror repositories, git doesn't allow you to push a single
refspec or tag.  Instead, it always pushes an exact copy of your current
repo, all reachable hashes and all tags.  When we have a mirror repo, we
only need to push it once, no matter how many tags we might be creating
releases for.

Signed-off-by: Doug Ledford <dledford@redhat.com>